LayoutTests/media/media-controller-play-then-pause-expected.txt


RUN(controller = video.controller)
EVENT(canplaythrough)
RUN(controller.play())
EVENT(playing)
RUN(controller.pause())
EVENT(pause)
RUN(controller.play())
EVENT(play)
EXPECTED (controller.playbackState == 'playing') OK
END OF TEST
